♻️ refactor : 스터디 노트 목록/생성 뷰 통합
